+++ crates/ethcore/src/client/test_client.rs
@@ -1247,6 +1247,8 @@ impl super::traits::EngineClient for TestBlockChainClient {
         self.miner
             .create_pending_block_at(self, txns, timestamp, block_number)
     }
+
+    fn demand_shutdown(&self) {}
 }
 
 impl PrometheusMetrics for TestBlockChainClient {